Refactor SPI package: Remove deprecated do.go, update connection methods, and enhance error handling
- Deleted the `do.go` file, consolidating functionality into other components.
- Updated connection methods in `machsvr_test.go`, `metrics.go`, and `watch.go` to use the new `spi.Connect` method instead of the previous `api.WithPassword` approach.
- Refactored `ShowTable` and `ShowTables` functions to utilize the new `TableDescription` type.
- Improved error handling in `metrics.go` to log warnings for metrics writing with no affected rows.
- Removed the `sql_wrap.go` file and its associated wrapped connection logic, simplifying the database interaction layer.
- Updated unit tests to reflect changes in the connection methods and removed obsolete tests related to the wrapped SQL result and row handling.
