+++ mods/server/http_query.go
@@ -307,7 +307,7 @@ func (svr *httpd) handleFileQuery(ctx *gin.Context) {
 		sqlParams = append(sqlParams, ts.Add(3*time.Second).UnixNano())
 		sqlParams = append(sqlParams, fileID)
 	} else if tableType == api.TableTypeTag {
-		var desc *api.TableDescription
+		var desc *spi.TableDescription
 		if rs := spi.ShowTable(ctx, conn, "MACHBASEDB", "SYS", tableName, false); rs.Err() != nil {
 			err = rs.Err()
 			rsp.Reason = fmt.Sprintf("fail to get table info '%s', %s", tableName, err.Error())
@@ -499,7 +499,7 @@ func (svr *httpd) handleTags(ctx *gin.Context) {
 		isCancelled = true
 	}()
 
-	var desc *api.TableDescription
+	var desc *spi.TableDescription
 	if rs := spi.ShowTable(ctx, conn, "MACHBASEDB", "SYS", table, false); rs.Err() != nil {
 		rsp.Success, rsp.Reason = false, rs.Err().Error()
 		rsp.Elapse = time.Since(tick).String()
@@ -569,7 +569,7 @@ func (svr *httpd) handleTagStat(ctx *gin.Context) {
 	}
 	defer conn.Close()
 
-	var desc *api.TableDescription
+	var desc *spi.TableDescription
 	if rs := spi.ShowTable(ctx, conn, "MACHBASEDB", "SYS", table, false); rs.Err() != nil {
 		rsp.Success, rsp.Reason = false, rs.Err().Error()
 		rsp.Elapse = time.Since(tick).String()
