+++ mods/server/http_write.go
@@ -87,7 +87,7 @@ func (svr *httpd) handleWrite(ctx *gin.Context) {
 	}
 	defer conn.Close()
 
-	var desc *api.TableDescription
+	var desc *spi.TableDescription
 	if rs := spi.ShowTable(ctx, conn, "MACHBASEDB", "SYS", tableName, false); rs.Err() != nil {
 		errRsp(http.StatusInternalServerError, fmt.Sprintf("fail to get table info '%s', %s", tableName, rs.Err().Error()))
 		return
@@ -294,7 +294,7 @@ func (svr *httpd) handleFileWrite(ctx *gin.Context) {
 		return
 	}
 
-	var desc *api.TableDescription
+	var desc *spi.TableDescription
 	if rs := spi.ShowTable(ctx, conn, "MACHBASEDB", "SYS", tableName, false); rs.Err() != nil {
 		err = rs.Err()
 		rsp.Reason = fmt.Sprintf("fail to get table info '%s', %s", tableName, err.Error())
@@ -485,7 +485,7 @@ func (svr *httpd) handleLineWrite(ctx *gin.Context) {
 	defer conn.Close()
 
 	dbName := ctx.Query("db")
-	var desc *api.TableDescription
+	var desc *spi.TableDescription
 	if rs := spi.ShowTable(ctx, conn, "MACHBASEDB", "SYS", dbName, false); rs.Err() != nil {
 		ctx.JSON(
 			http.StatusBadRequest,
