+++ spi/database.go
@@ -12,6 +12,7 @@ import (
 	"errors"
 	"fmt"
 	"net"
+	"os"
 	"strconv"
 	"strings"
 	"sync"
@@ -453,7 +454,8 @@ func MakeBuffer(columnTypes []*sql.ColumnType) []interface{} {
 			case "DATE", "DATETIME", "TIMESTAMP":
 				buffer[i] = new(sql.NullTime)
 			default:
-				fmt.Println("=================>", colType.DatabaseTypeName(), colType.ScanType().String())
+				fmt.Printf("=================> colName: %s, databaseType: %s, scanType: %s\n",
+					colType.Name(), colType.DatabaseTypeName(), colType.ScanType().String())
 				buffer[i] = new(interface{})
 			}
 		}
@@ -578,4 +580,613 @@ func ColumnTypes(rows *sql.Rows) ([]string, error) {
 		}
 	}
 	return types, nil
+}
+
+type LicenseInfo struct {
+	Id            string `json:"id"`
+	Type          string `json:"type"`
+	Customer      string `json:"customer"`
+	Project       string `json:"project"`
+	CountryCode   string `json:"countryCode"`
+	InstallDate   string `json:"installDate"`
+	IssueDate     string `json:"issueDate"`
+	LicenseStatus string `json:"licenseStatus,omitempty"`
+}
+
+func GetLicenseInfo(ctx context.Context, conn *sql.Conn) (*LicenseInfo, error) {
+	ret := &LicenseInfo{}
+	var violateStatus int
+	var violateMsg sql.NullString
+	row := conn.QueryRowContext(ctx, "select ID, TYPE, CUSTOMER, PROJECT, COUNTRY_CODE, INSTALL_DATE, ISSUE_DATE, VIOLATE_STATUS, VIOLATE_MSG from v$license_info")
+	if err := row.Err(); err != nil {
+		return nil, err
+	}
+	if err := row.Scan(&ret.Id, &ret.Type, &ret.Customer, &ret.Project, &ret.CountryCode, &ret.InstallDate, &ret.IssueDate, &violateStatus, &violateMsg); err != nil {
+		return nil, err
+	}
+	if violateStatus == 0 {
+		ret.LicenseStatus = "Valid"
+	} else if violateMsg.Valid {
+		ret.LicenseStatus = violateMsg.String
+	}
+	return ret, nil
+}
+
+func InstallLicenseFile(ctx context.Context, conn *sql.Conn, path string) (*LicenseInfo, error) {
+	if strings.ContainsRune(path, ';') {
+		return nil, errors.New("invalid license file path")
+	}
+	_, err := conn.ExecContext(ctx, "alter system install license='"+path+"'")
+	if err != nil {
+		return nil, err
+	}
+	return GetLicenseInfo(ctx, conn)
+}
+
+func InstallLicenseData(ctx context.Context, conn *sql.Conn, licenseFilePath string, content []byte) (*LicenseInfo, error) {
+	_, err := os.Stat(licenseFilePath)
+	if err == nil {
+		// backup existing file
+		os.Rename(licenseFilePath, fmt.Sprintf("%s_%s", licenseFilePath, time.Now().Format("20060102_150405")))
+	}
+	if err := os.WriteFile(licenseFilePath, content, 0640); err != nil {
+		return nil, err
+	}
+	return InstallLicenseFile(ctx, conn, licenseFilePath)
+}
+
+type TableInfo struct {
+	Database string        `json:"database"`       // M$SYS_TABLES.DATABASE_ID
+	User     string        `json:"user"`           // M$SYS_USERS.NAME
+	Name     string        `json:"name"`           // M$SYS_TABLES.NAME
+	Id       int64         `json:"id"`             // M$SYS_TABLES.ID
+	Type     api.TableType `json:"type"`           // M$SYS_TABLES.TYPE
+	Flag     api.TableFlag `json:"flag,omitempty"` // M$SYS_TABLES.FLAG
+}
+
+func (ti *TableInfo) Kind() string {
+	desc := "undef"
+	switch ti.Type {
+	case api.TableTypeLog:
+		desc = "Log Table"
+	case api.TableTypeFixed:
+		desc = "Fixed Table"
+	case api.TableTypeVolatile:
+		desc = "Volatile Table"
+	case api.TableTypeLookup:
+		desc = "Lookup Table"
+	case api.TableTypeKeyValue:
+		desc = "KeyValue Table"
+	case api.TableTypeTag:
+		desc = "Tag Table"
+	}
+	switch ti.Flag {
+	case api.TableFlagData:
+		desc += " (data)"
+	case api.TableFlagRollup:
+		desc += " (rollup)"
+	case api.TableFlagMeta:
+		desc += " (meta)"
+	case api.TableFlagStat:
+		desc += " (stat)"
+	}
+	return desc
+}
+
+func ifThenElse(cond bool, a, b string) string {
+	if cond {
+		return a
+	}
+	return b
+}
+
+func ListTablesWalk(ctx context.Context, conn *sql.Conn, showAll bool, callback func(*TableInfo, error) bool) {
+	descriptiveType := false
+	sqlText := SqlTidy(
+		`SELECT
+			j.DB_NAME as DATABASE_NAME,
+			u.NAME as USER_NAME,
+			j.NAME as TABLE_NAME,
+			j.ID as TABLE_ID,`,
+		ifThenElse(descriptiveType, `
+			case j.TYPE
+				when 0 then 'Log'
+				when 1 then 'Fixed'
+				when 3 then 'Volatile'
+				when 4 then 'Lookup'
+				when 5 then 'KeyValue'
+				when 6 then 'Tag'
+				else ''
+			end as TABLE_TYPE,
+			case j.FLAG
+				when 1 then 'Data'
+				when 2 then 'Rollup'
+				when 4 then 'Meta'
+				when 8 then 'Stat'
+				else ''
+			end as TABLE_FLAG`,
+			`
+			j.TYPE as TABLE_TYPE,
+			j.FLAG as TABLE_FLAG`),
+		`FROM
+			M$SYS_USERS u,
+			(
+				select
+					a.ID as ID,
+					a.NAME as NAME,
+					a.USER_ID as USER_ID,
+					a.TYPE as TYPE,
+					a.FLAG as FLAG,
+					case a.DATABASE_ID
+						when -1 then 'MACHBASEDB'
+						else d.MOUNTDB
+					end as DB_NAME
+				from
+					M$SYS_TABLES a
+				left join
+					V$STORAGE_MOUNT_DATABASES d
+				on
+					a.DATABASE_ID = d.BACKUP_TBSID
+			) as j
+		WHERE
+			u.USER_ID = j.USER_ID`,
+		ifThenElse(showAll, "", "AND SUBSTR(j.NAME, 1, 1) <> '_'"),
+		`ORDER by j.NAME`)
+	rows, err := conn.QueryContext(ctx, sqlText)
+	if err != nil {
+		callback(nil, err)
+		return
+	}
+	defer rows.Close()
+
+	for rows.Next() {
+		ti := &TableInfo{}
+		err = rows.Scan(&ti.Database, &ti.User, &ti.Name, &ti.Id, &ti.Type, &ti.Flag)
+		if !callback(ti, err) {
+			return
+		}
+	}
+	if err := rows.Err(); err != nil {
+		callback(nil, err)
+	}
+}
+
+func QueryTableType(ctx context.Context, conn *sql.Conn, fullTableName string) (api.TableType, error) {
+	_, userName, tableName := api.TableName(fullTableName).Split()
+	sql := "select type from M$SYS_TABLES T, M$SYS_USERS U where U.NAME = ? and U.USER_ID = T.USER_ID AND T.NAME = ?"
+	r := conn.QueryRowContext(ctx, sql, strings.ToUpper(userName), strings.ToUpper(tableName))
+	if r.Err() != nil {
+		return -1, r.Err()
+	}
+	var ret api.TableType
+	if err := r.Scan(&ret); err != nil {
+		return -1, err
+	}
+	return ret, nil
+}
+
+func TruncateTableIfExists(ctx context.Context, conn *sql.Conn, fullTableName string, truncate bool) (exists bool, truncated bool, err error) {
+	exists, err = ExistsTable(ctx, conn, fullTableName)
+	if err != nil {
+		return
+	}
+	if !exists {
+		return
+	}
+
+	// TRUNCATE TABLE
+	if !truncate {
+		return
+	}
+	tableType, err0 := QueryTableType(ctx, conn, fullTableName)
+	if err0 != nil {
+		err = fmt.Errorf("table '%s' doesn't exist, %s", fullTableName, err0.Error())
+		return
+	}
+	if tableType == api.TableTypeLog {
+		_, err0 := conn.ExecContext(ctx, fmt.Sprintf("truncate table %s", fullTableName))
+		if err0 != nil {
+			err = err0
+			return
+		}
+		truncated = true
+	} else {
+		_, err0 := conn.ExecContext(ctx, fmt.Sprintf("delete from %s", fullTableName))
+		if err0 != nil {
+			err = err0
+			return
+		}
+		truncated = true
+	}
+	return
+}
+
+func ExistsTable(ctx context.Context, conn *sql.Conn, fullTableName string) (bool, error) {
+	_, userName, tableName := api.TableName(fullTableName).Split()
+	sql := "select count(*) from M$SYS_TABLES T, M$SYS_USERS U where U.NAME = ? and U.USER_ID = T.USER_ID AND T.NAME = ?"
+	r := conn.QueryRowContext(ctx, sql, strings.ToUpper(userName), strings.ToUpper(tableName))
+	if err := r.Err(); err != nil {
+		fmt.Println("error", err.Error())
+		return false, err
+	}
+	var count = 0
+	if err := r.Scan(&count); err != nil {
+		return false, err
+	}
+	return (count == 1), nil
+}
+
+// TableDescription is represents data that comes as a result of 'desc <table>'
+type TableDescription struct {
+	Database string              `json:"database"`
+	User     string              `json:"user"`
+	Name     string              `json:"name"`
+	Id       int64               `json:"id"`
+	Type     api.TableType       `json:"type"`
+	Flag     api.TableFlag       `json:"flag,omitempty"`
+	Columns  api.Columns         `json:"columns"`
+	Indexes  []*IndexDescription `json:"indexes"`
+
+	Summarized       bool   `json:"summarized"`
+	SummarizedColumn string `json:"summarizedColum,omitempty"`
+	TagNameColumn    string `json:"tagNameColumn,omitempty"`
+}
+
+type IndexDescription struct {
+	Id             int64         `json:"id"`
+	Name           string        `json:"name"`
+	Type           api.IndexType `json:"type"`
+	Cols           []string      `json:"columns"`
+	KeyCompress    bool          `json:"keyCompress"`
+	MaxLevel       int           `json:"maxLevel"`
+	PartValueCount int           `json:"partValueCount"`
+	BitMapEncode   string        `json:"bitMapEncode"`
+}
+
+// String returns string representation of table type.
+func (td *TableDescription) String() string {
+	desc := "undef"
+	switch td.Type {
+	case api.TableTypeLog:
+		desc = "Log Table"
+	case api.TableTypeFixed:
+		desc = "Fixed Table"
+	case api.TableTypeVolatile:
+		desc = "Volatile Table"
+	case api.TableTypeLookup:
+		desc = "Lookup Table"
+	case api.TableTypeKeyValue:
+		desc = "KeyValue Table"
+	case api.TableTypeTag:
+		desc = "Tag Table"
+	}
+	switch td.Flag {
+	case api.TableFlagData:
+		desc += " (data)"
+	case api.TableFlagRollup:
+		desc += " (rollup)"
+	case api.TableFlagMeta:
+		desc += " (meta)"
+	case api.TableFlagStat:
+		desc += " (stat)"
+	}
+	return desc
+}
+
+// Describe retrieves the result of 'desc table'.
+//
+// If includeHiddenColumns is true, the result includes hidden columns those name start with '_'
+// such as "_RID" and "_ARRIVAL_TIME".
+func DescribeTable(ctx context.Context, conn *sql.Conn, name string, includeHiddenColumns bool) (*TableDescription, error) {
+	_, _, tableName := api.TableName(name).Split()
+	if strings.HasPrefix(tableName, "V$") {
+		return describe_mv(ctx, conn, api.TableName(name), includeHiddenColumns)
+	} else if strings.HasPrefix(tableName, "M$") {
+		return describe_mv(ctx, conn, api.TableName(name), includeHiddenColumns)
+	} else {
+		return describe(ctx, conn, api.TableName(name), includeHiddenColumns)
+	}
+}
+
+func describe(ctx context.Context, conn *sql.Conn, name api.TableName, includeHiddenColumns bool) (*TableDescription, error) {
+	d := &TableDescription{}
+	var colCount int
+
+	dbName, userName, tableName := name.Split()
+	dbId := -1
+
+	if dbName != "" && dbName != "MACHBASEDB" {
+		row := conn.QueryRowContext(ctx, "select BACKUP_TBSID from V$STORAGE_MOUNT_DATABASES where MOUNTDB = ?", dbName)
+		if err := row.Err(); err != nil {
+			return nil, err
+		}
+		if err := row.Scan(&dbId); err != nil {
+			return nil, err
+		}
+	}
+
+	describeSqlText := SqlTidy(
+		`SELECT
+			j.ID as TABLE_ID,
+			j.TYPE as TABLE_TYPE,
+			j.FLAG as TABLE_FLAG,
+			j.COLCOUNT as TABLE_COLCOUNT
+		from
+			M$SYS_USERS u,
+			M$SYS_TABLES j
+		where
+			u.NAME = ?
+		and j.USER_ID = u.USER_ID
+		and j.DATABASE_ID = ?
+		and j.NAME = ?`)
+
+	r := conn.QueryRowContext(ctx, describeSqlText, userName, dbId, tableName)
+	if err := r.Err(); err != nil {
+		return nil, err
+	}
+	if err := r.Scan(&d.Id, &d.Type, &d.Flag, &colCount); err != nil {
+		return nil, err
+	}
+	d.Database = dbName
+	d.User = userName
+	d.Name = tableName
+
+	rows, err := conn.QueryContext(ctx, "select name, type, length, id, flag from M$SYS_COLUMNS where table_id = ? AND database_id = ? order by id", d.Id, dbId)
+	if err != nil {
+		return nil, err
+	}
+	defer func() {
+		if rows != nil {
+			rows.Close()
+		}
+	}()
+
+	for rows.Next() {
+		col := &api.Column{}
+		err = rows.Scan(&col.Name, &col.Type, &col.Length, &col.Id, &col.Flag)
+		if err != nil {
+			return nil, err
+		}
+		if !includeHiddenColumns && strings.HasPrefix(col.Name, "_") {
+			continue
+		}
+		col.DataType = col.Type.DataType()
+		d.Columns = append(d.Columns, col)
+
+		if col.Flag&api.ColumnFlagSummarized > 0 {
+			d.Summarized = true
+			d.SummarizedColumn = col.Name
+		}
+		if col.Flag&api.ColumnFlagTagName > 0 {
+			d.TagNameColumn = col.Name
+		}
+		if col.Flag&api.ColumnFlagBasetime > 0 && col.Type != api.ColumnTypeDatetime {
+			col.Flag = api.ColumnFlagBaseDistance
+		}
+	}
+	rows.Close()
+	rows = nil
+
+	if indexes, err := describe_idx(ctx, conn, d.Id, dbId); err != nil {
+		return nil, err
+	} else {
+		d.Indexes = indexes
+	}
+	return d, nil
+}
+
+func describe_mv(ctx context.Context, conn *sql.Conn, name api.TableName, includeHiddenColumns bool) (*TableDescription, error) {
+	d := &TableDescription{}
+	var tableType int
+	var colCount int
+
+	d.Database, d.User, d.Name = name.Split()
+	tablesTable := "M$SYS_TABLES"
+	columnsTable := "M$SYS_COLUMNS"
+	if strings.HasPrefix(d.Name, "V$") {
+		tablesTable = "V$TABLES"
+		columnsTable = "V$COLUMNS"
+	} else if strings.HasPrefix(d.Name, "M$") {
+		tablesTable = "M$TABLES"
+		columnsTable = "M$COLUMNS"
+	}
+	r := conn.QueryRowContext(ctx, fmt.Sprintf("select name, type, flag, id, colcount from %s where name = ?", tablesTable), d.Name)
+	if err := r.Err(); err != nil {
+		return nil, err
+	}
+	if err := r.Scan(&d.Name, &tableType, &d.Flag, &d.Id, &colCount); err != nil {
+		return nil, err
+	}
+	d.Type = api.TableType(tableType)
+
+	rows, err := conn.QueryContext(ctx, fmt.Sprintf(`select name, type, length, id from %s where table_id = ? order by id`, columnsTable), d.Id)
+	if err != nil {
+		return nil, err
+	}
+	defer rows.Close()
+
+	for rows.Next() {
+		col := &api.Column{}
+		err = rows.Scan(&col.Name, &col.Type, &col.Length, &col.Id)
+		if err != nil {
+			return nil, err
+		}
+		if !includeHiddenColumns && strings.HasPrefix(col.Name, "_") {
+			continue
+		}
+		col.DataType = col.Type.DataType()
+		d.Columns = append(d.Columns, col)
+	}
+	return d, nil
+}
+
+func describe_idx(ctx context.Context, conn *sql.Conn, tableId int64, dbId int) ([]*IndexDescription, error) {
+	rows, err := conn.QueryContext(ctx,
+		`select
+			b.name,
+			b.type,
+			b.id,
+			b.key_compress,
+			b.max_level,
+			b.part_value_count,
+			case b.bitmap_encode
+				when 0 then 'EQUAL'
+				else 'RANGE' end 
+			as bitmap_encode
+		from
+			M$SYS_TABLES  a,
+			M$SYS_INDEXES b
+		where
+			a.id = ?
+		AND a.database_id = ?
+		AND a.id = b.table_id
+		`, tableId, dbId)
+	if err != nil {
+		return nil, err
+	}
+	defer rows.Close()
+
+	indexes := []*IndexDescription{}
+	for rows.Next() {
+		d := &IndexDescription{}
+		var indexType int
+		var keyCompress int
+		if err = rows.Scan(&d.Name, &indexType, &d.Id, &keyCompress, &d.MaxLevel, &d.PartValueCount, &d.BitMapEncode); err != nil {
+			return nil, err
+		}
+		d.Type = api.IndexType(indexType)
+		d.KeyCompress = (keyCompress != 0)
+		idxCols, err := conn.QueryContext(ctx, `select name from M$SYS_INDEX_COLUMNS where index_id = ? AND database_id = ? order by col_id`, d.Id, dbId)
+		if err != nil {
+			return nil, err
+		}
+		for idxCols.Next() {
+			var col string
+			if err = idxCols.Scan(&col); err != nil {
+				idxCols.Close()
+				return nil, err
+			}
+			d.Cols = append(d.Cols, col)
+		}
+		idxCols.Close()
+		indexes = append(indexes, d)
+	}
+	return indexes, nil
+}
+
+/* Interpreting Influx line protocol
+
+   | Machbase            | influxdb                                    |
+   | ------------------- | ------------------------------------------- |
+   | table name          | db                                          |
+   | tag name            | measurement + '.' + field name              |
+   | time                | timestamp                                   |
+   | value               | value of the field (if it is not a number type, will be ignored and not inserted) |
+*/
+
+func WriteLineProtocol(ctx context.Context, conn *sql.Conn, dbName string, descColumns api.Columns, measurement string, fields map[string]any, tags map[string]string, ts time.Time) api.Result {
+	columns := descColumns.Names()
+	columns = columns[:3]
+
+	/*
+		Machbase : name, time, value, host
+		influxdb : tags key[DC, HOST, NAME, SYSTEM]
+		=> HOST append / DC, NAME, SYSTEM not append
+	*/
+	compareNames := descColumns.Names()
+	compareTypes := descColumns.DataTypes()
+	compareNames = compareNames[3:]
+	compareTypes = compareTypes[3:]
+	for idx, val := range compareNames {
+		if _, ok := tags[val]; ok {
+			if compareTypes[idx] == api.DataTypeString {
+				columns = append(columns, val)
+			}
+		}
+	}
+
+	rows := make([][]any, 0)
+
+	for k, v := range fields {
+		values := make([]any, 0)
+		values = append(values, fmt.Sprintf("%s.%s", measurement, k))
+		values = append(values, ts)
+
+		switch val := v.(type) {
+		case float32:
+			values = append(values, float64(val))
+		case float64:
+			values = append(values, val)
+		case int:
+			values = append(values, float64(val))
+		case int32:
+			values = append(values, float64(val))
+		case int64:
+			values = append(values, float64(val))
+		default:
+			// unsupported value type
+			continue
+		}
+
+		for i := 3; i < len(columns); i++ {
+			values = append(values, tags[columns[i]])
+		}
+
+		rows = append(rows, values)
+	}
+
+	if len(rows) == 0 {
+		return &InsertResult{
+			rowsAffected: 0,
+			message:      "no rows inserted",
+		}
+	}
+
+	vf := make([]string, len(columns))
+	for i := range vf {
+		vf[i] = "?"
+	}
+	tableName := dbName
+	valuesPlaces := strings.Join(vf, ",")
+	columnsPhrase := strings.Join(columns, ",")
+
+	sqlText := fmt.Sprintf("INSERT INTO %s(%s) VALUES(%s)", tableName, columnsPhrase, valuesPlaces)
+	var numRows int
+	for _, rec := range rows {
+		_, err := conn.ExecContext(ctx, sqlText, rec...)
+		if err != nil {
+			return &InsertResult{
+				err:          err,
+				rowsAffected: numRows,
+				message:      "batch inserts aborted - " + sqlText,
+			}
+		}
+		numRows++
+	}
+
+	ret := &InsertResult{
+		rowsAffected: numRows,
+	}
+	ret.message = MakeUserMessage(SQLStatementTypeInsert, int64(numRows))
+	return ret
+}
+
+var _ api.Result = &InsertResult{}
+
+type InsertResult struct {
+	err          error
+	rowsAffected int
+	message      string
+}
+
+func (ir *InsertResult) Err() error {
+	return ir.err
+}
+
+func (ir *InsertResult) RowsAffected() int64 {
+	return int64(ir.rowsAffected)
+}
+
+func (ir *InsertResult) Message() string {
+	return ir.message
 }
