+++ spi/watch.go
@@ -133,7 +133,7 @@ func (w *Watcher) init(ctx context.Context) error {
 		}
 	}
 
-	var desc *api.TableDescription
+	var desc *TableDescription
 	if rs := ShowTable(ctx, conn, "", "", w.TableName, false); rs.err != nil {
 		return fmt.Errorf("fail to get table info '%s', %s", w.TableName, rs.err.Error())
 	} else {
